pyproject.toml
[project]
name = "web-reputational-analysis"
version = "0.5.0"
description = "Pipeline per la raccolta, normalizzazione ed esportazione di dati da fonti web per web reputational analysis."
requires-python = ">=3.11"
dependencies = [
    "requests>=2.28",
    "python-dateutil>=2.8",
    "python-dotenv>=1.0",
    "Wikipedia-API>=0.7",
]

[project.optional-dependencies]
# Enrichment NLP: language detection + sentiment analysis multilingue.
# Installare con: pip install ".[nlp]"
# Nota: torch richiede ~200 MB (CPU-only). Il modello XLM-RoBERTa viene
# scaricato automaticamente da HuggingFace alla prima esecuzione (~1.1 GB).
nlp = [
    "langdetect>=1.0.9",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
    "tiktoken>=0.5.0",
]
dev = ["pytest>=8.0"]

[project.scripts]
wra = "main:main"

[tool.setuptools.packages.find]
# Esclude directory non-package: data/ (storage file), tests/, docs/
exclude = ["data*", "tests*", "docs*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
# Aggiunge la root del progetto a sys.path: evita sys.path.insert in main.py
# e permette ai test di importare i moduli direttamente.
pythonpath = ["."]
testpaths  = ["tests"]
